+++ decentraland/components/presentational/LinkToBlockchain.test.js
